Change decimal to float

+++ Collaborative Resource Management System/Collaborative Resource Management System/Models/Consumable.cs
@@ -6,7 +6,7 @@ namespace Collaborative_Resource_Management_System.Models
     {
         public int ConsumableID { get; set; }
         public int ItemID { get; set; }
-        public decimal PricePerUnit { get; set; }
+        public float PricePerUnit { get; set; }
         public int QuantityAvailable { get; set; }
         public int MinimumQuantity { get; set; }
     }
